--- app/client_secret.json
@@ -1,1 +1,0 @@
-{"web":{"client_id":"782606518020-j8sf7ke5879ivei5p5anolfmls93pblq.apps.googleusercontent.com","project_id":"youtubereplay-project","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"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","redirect_uris":["http://localhost:8080"],"javascript_origins":["http://localhost:8080"]}}
