+++ app/config/definitions.py
@@ -1,0 +1,2 @@
+import os
+ROOT_DIR = os.path.realpath(os.path.join(os.path.dirname(__file__), '..'))
